+++ src/set-operations.cpp
@@ -9,9 +9,9 @@ using namespace Rcpp;
 
 // [[Rcpp::export]]
 List phint_phint_intersect_cpp(
-    IntegerVector x_size, List x_starts, List x_ends,
-    IntegerVector y_size, List y_starts, List y_ends,
-    String bounds
+  IntegerVector x_size, List x_starts, List x_ends,
+  IntegerVector y_size, List y_starts, List y_ends,
+  String bounds
 ) {
   PhintVectorView x {x_size, x_starts, x_ends};
   PhintVectorView y {y_size, y_starts, y_ends};
@@ -24,9 +24,9 @@ List phint_phint_intersect_cpp(
 
 // [[Rcpp::export]]
 List phint_intvl_intersect_cpp(
-    IntegerVector x_size, List x_starts, List x_ends,
-    DatetimeVector y_starts, NumericVector y_spans,
-    String bounds
+  IntegerVector x_size, List x_starts, List x_ends,
+  DatetimeVector y_starts, NumericVector y_spans,
+  String bounds
 ) {
   PhintVectorView x {x_size, x_starts, x_ends};
   IntvlVectorView y {y_starts, y_spans};
@@ -39,9 +39,9 @@ List phint_intvl_intersect_cpp(
 
 // [[Rcpp::export]]
 List intvl_phint_intersect_cpp(
-    DatetimeVector x_starts, NumericVector x_spans,
-    IntegerVector y_size, List y_starts, List y_ends,
-    String bounds
+  DatetimeVector x_starts, NumericVector x_spans,
+  IntegerVector y_size, List y_starts, List y_ends,
+  String bounds
 ) {
   IntvlVectorView x {x_starts, x_spans};
   PhintVectorView y {y_size, y_starts, y_ends};
@@ -54,9 +54,9 @@ List intvl_phint_intersect_cpp(
 
 // [[Rcpp::export]]
 List intvl_intvl_intersect_cpp(
-    DatetimeVector x_starts, NumericVector x_spans,
-    DatetimeVector y_starts, NumericVector y_spans,
-    String bounds
+  DatetimeVector x_starts, NumericVector x_spans,
+  DatetimeVector y_starts, NumericVector y_spans,
+  String bounds
 ) {
   IntvlVectorView x {x_starts, x_spans};
   IntvlVectorView y {y_starts, y_spans};
@@ -71,8 +71,8 @@ List intvl_intvl_intersect_cpp(
 
 // [[Rcpp::export]]
 List phint_phint_union_cpp(
-    IntegerVector x_size, List x_starts, List x_ends,
-    IntegerVector y_size, List y_starts, List y_ends
+  IntegerVector x_size, List x_starts, List x_ends,
+  IntegerVector y_size, List y_starts, List y_ends
 ) {
   return phint_operate(
     PhintVectorView{x_size, x_starts, x_ends},
@@ -83,8 +83,8 @@ List phint_phint_union_cpp(
 
 // [[Rcpp::export]]
 List phint_intvl_union_cpp(
-    IntegerVector x_size, List x_starts, List x_ends,
-    DatetimeVector y_starts, NumericVector y_spans
+  IntegerVector x_size, List x_starts, List x_ends,
+  DatetimeVector y_starts, NumericVector y_spans
 ) {
   return phint_operate(
     PhintVectorView{x_size, x_starts, x_ends},
@@ -95,8 +95,8 @@ List phint_intvl_union_cpp(
 
 // [[Rcpp::export]]
 List intvl_phint_union_cpp(
-    DatetimeVector x_starts, NumericVector x_spans,
-    IntegerVector y_size, List y_starts, List y_ends
+  DatetimeVector x_starts, NumericVector x_spans,
+  IntegerVector y_size, List y_starts, List y_ends
 ) {
   return phint_operate(
     IntvlVectorView{x_starts, x_spans},
@@ -107,8 +107,8 @@ List intvl_phint_union_cpp(
 
 // [[Rcpp::export]]
 List intvl_intvl_union_cpp(
-    DatetimeVector x_starts, NumericVector x_spans,
-    DatetimeVector y_starts, NumericVector y_spans
+  DatetimeVector x_starts, NumericVector x_spans,
+  DatetimeVector y_starts, NumericVector y_spans
 ) {
   return phint_operate(
     IntvlVectorView{x_starts, x_spans},
@@ -121,8 +121,8 @@ List intvl_intvl_union_cpp(
 
 // [[Rcpp::export]]
 List phint_phint_setdiff_cpp(
-    IntegerVector x_size, List x_starts, List x_ends,
-    IntegerVector y_size, List y_starts, List y_ends
+  IntegerVector x_size, List x_starts, List x_ends,
+  IntegerVector y_size, List y_starts, List y_ends
 ) {
   return phint_operate(
     PhintVectorView{x_size, x_starts, x_ends},
@@ -133,8 +133,8 @@ List phint_phint_setdiff_cpp(
 
 // [[Rcpp::export]]
 List phint_intvl_setdiff_cpp(
-    IntegerVector x_size, List x_starts, List x_ends,
-    DatetimeVector y_starts, NumericVector y_spans
+  IntegerVector x_size, List x_starts, List x_ends,
+  DatetimeVector y_starts, NumericVector y_spans
 ) {
   return phint_operate(
     PhintVectorView{x_size, x_starts, x_ends},
@@ -145,8 +145,8 @@ List phint_intvl_setdiff_cpp(
 
 // [[Rcpp::export]]
 List intvl_phint_setdiff_cpp(
-    DatetimeVector x_starts, NumericVector x_spans,
-    IntegerVector y_size, List y_starts, List y_ends
+  DatetimeVector x_starts, NumericVector x_spans,
+  IntegerVector y_size, List y_starts, List y_ends
 ) {
   return phint_operate(
     IntvlVectorView{x_starts, x_spans},
@@ -157,8 +157,8 @@ List intvl_phint_setdiff_cpp(
 
 // [[Rcpp::export]]
 List intvl_intvl_setdiff_cpp(
-    DatetimeVector x_starts, NumericVector x_spans,
-    DatetimeVector y_starts, NumericVector y_spans
+  DatetimeVector x_starts, NumericVector x_spans,
+  DatetimeVector y_starts, NumericVector y_spans
 ) {
   return phint_operate(
     IntvlVectorView{x_starts, x_spans},
@@ -171,8 +171,8 @@ List intvl_intvl_setdiff_cpp(
 
 // [[Rcpp::export]]
 List phint_phint_symmetric_setdiff_cpp(
-    IntegerVector x_size, List x_starts, List x_ends,
-    IntegerVector y_size, List y_starts, List y_ends
+  IntegerVector x_size, List x_starts, List x_ends,
+  IntegerVector y_size, List y_starts, List y_ends
 ) {
   return phint_operate(
     PhintVectorView{x_size, x_starts, x_ends},
@@ -183,8 +183,8 @@ List phint_phint_symmetric_setdiff_cpp(
 
 // [[Rcpp::export]]
 List phint_intvl_symmetric_setdiff_cpp(
-    IntegerVector x_size, List x_starts, List x_ends,
-    DatetimeVector y_starts, NumericVector y_spans
+  IntegerVector x_size, List x_starts, List x_ends,
+  DatetimeVector y_starts, NumericVector y_spans
 ) {
   return phint_operate(
     PhintVectorView{x_size, x_starts, x_ends},
@@ -195,8 +195,8 @@ List phint_intvl_symmetric_setdiff_cpp(
 
 // [[Rcpp::export]]
 List intvl_phint_symmetric_setdiff_cpp(
-    DatetimeVector x_starts, NumericVector x_spans,
-    IntegerVector y_size, List y_starts, List y_ends
+  DatetimeVector x_starts, NumericVector x_spans,
+  IntegerVector y_size, List y_starts, List y_ends
 ) {
   return phint_operate(
     IntvlVectorView{x_starts, x_spans},
@@ -207,8 +207,8 @@ List intvl_phint_symmetric_setdiff_cpp(
 
 // [[Rcpp::export]]
 List intvl_intvl_symmetric_setdiff_cpp(
-    DatetimeVector x_starts, NumericVector x_spans,
-    DatetimeVector y_starts, NumericVector y_spans
+  DatetimeVector x_starts, NumericVector x_spans,
+  DatetimeVector y_starts, NumericVector y_spans
 ) {
   return phint_operate(
     IntvlVectorView{x_starts, x_spans},